+++ packages/firebase_auth/lib/firebase_auth.dart
@@ -434,6 +434,21 @@ class FirebaseAuth {
     return currentUser;
   }
 
+  Future<FirebaseUser> linkWithTwitterCredential({
+    @required String authToken,
+    @required String authTokenSecret,
+  }) async {
+    final Map<dynamic, dynamic> data = await channel.invokeMethod(
+      'linkWithTwitterCredential',
+      <String, String>{
+        'authToken': authToken,
+        'authTokenSecret': authTokenSecret,
+      },
+    );
+    final FirebaseUser currentUser = FirebaseUser._(data);
+    return currentUser;
+  }
+
   /// Sets the user-facing language code for auth operations that can be
   /// internationalized, such as [sendEmailVerification]. This language
   /// code should follow the conventions defined by the IETF in BCP47.